lib: refacto rename SCLang to SynthDefMgt + rename to experimentPythonbindings
