+++ renardo_lib/renardo_lib/Midi.py
@@ -1,7 +1,7 @@
 """ Module for converting handling MIDI in/out and functions relating to MIDI pitch calculation. """
 
 
-from renardo_lib.SCLang import SynthDefProxy
+from renardo_lib.SynthDefManagement import SynthDefProxy
 
 try:
     import rtmidi
