--- renardo_lib/renardo_lib/SCLang/Effects/__init__.py
@@ -1,1 +1,0 @@
-from renardo_lib.SCLang.Effects.Util import *
