--- renardo_lib/renardo_lib/SCLang/__init__.py
@@ -1,5 +1,0 @@
-from renardo_lib.SCLang.SCLangExperimentalBindings import *
-from renardo_lib.SCLang.SynthDef import DefaultSynthDef
-from renardo_lib.SCLang.SynthDefProxy import SynthDefProxy
-from renardo_lib.SCLang.SampleSynthDefs import SampleSynthDef
-
