+++ renardo_lib/renardo_lib/SynthDefManagement/__init__.py
@@ -1,0 +1,5 @@
+from renardo_lib.SynthDefManagement.SCLangExperimentalPythonBindings import *
+from renardo_lib.SynthDefManagement.PySynthDef import DefaultSynthDef
+from renardo_lib.SynthDefManagement.SynthDefProxy import SynthDefProxy
+from renardo_lib.SynthDefManagement.SampleSynthDefs import SampleSynthDef
+
