+++ renardo_lib/renardo_lib/runtime/synthdefs_initialisation/sclang_file_synthdefs.py
@@ -1,4 +1,21 @@
-from renardo_lib.SCLang.FileSynthDef import FileSynthDef
+from renardo_lib.SynthDefManagement.FileSynthDef import FileSynthDef
+
+play = FileSynthDef('play1')
+play.defaults["dur"]=.5
+play.add()
+play = FileSynthDef('play2')
+play.defaults["dur"]=.5
+play.add()
+
+#loop = FileSynthDef('loop')
+#loop.defaults['pos'] = 0
+#loop.defaults['sample'] = 0
+## loop.defaults['spack'] = SAMPLES_PACK_NUMBER
+#loop.defaults['beat_stretch'] = 0
+#loop.add()
+
+#stretch = FileSynthDef('stretch')
+#stretch.add()
 
 # SynthDefs read from file
 bass = FileSynthDef('bass')
@@ -7,431 +24,431 @@ bass.add()
 dirt = FileSynthDef('dirt')
 dirt.add()
 
-sawbass = FileSynthDef('sawbass')
-sawbass.add()
-
-prophet = FileSynthDef('prophet')
-prophet.add()
-
-pads = FileSynthDef('pads')
-pads.add()
-
-pasha = FileSynthDef('pasha')
-pasha.add()
-
-ambi = FileSynthDef("ambi")
-ambi.add()
-
-space = FileSynthDef("space")
-space.add()
-
-keys = FileSynthDef("keys")
-keys.add()
-
-bass = FileSynthDef("bass")
-bass.add()
-
-dbass = FileSynthDef("dbass")
-dbass.add()
-
-sinepad = FileSynthDef("sinepad")
-sinepad.add()
-
-video = FileSynthDef("video")
-video.add()
-
-viola = FileSynthDef("viola")
-viola.add()
-
-hydra = FileSynthDef("hydra")
-hydra.add()
-
-abass = FileSynthDef("abass")
-abass.add()
-
-acidbass = FileSynthDef("acidbass")
-acidbass.add()
-
-alva = FileSynthDef("alva")
-alva.add()
-
-angel = FileSynthDef("angel")
-angel.add()
-
-angst = FileSynthDef("angst")
-angst.add()
-
-bassguitar = FileSynthDef("bassguitar")
-bassguitar.add()
-
-bbass = FileSynthDef("bbass")
-bbass.add()
-
-bchaos = FileSynthDef("bchaos")
-bchaos.add()
-
-bellmod = FileSynthDef("bellmod")
-bellmod.add()
-
-benoit = FileSynthDef("benoit")
-benoit.add()
-
-birdy = FileSynthDef("birdy")
-birdy.add()
-
-blips = FileSynthDef("blips")
-blips.add()
-
-bnoise = FileSynthDef("bnoise")
-bnoise.add()
-
-borgan = FileSynthDef("borgan")
-borgan.add()
-
-bounce = FileSynthDef("bounce")
-bounce.add()
-
-bphase = FileSynthDef("bphase")
-bphase.add()
-
-brass = FileSynthDef("brass")
-brass.add()
-
-brown = FileSynthDef("brown")
-brown.add()
-
-chimebell = FileSynthDef("chimebell")
-chimebell.add()
-
-chipsy = FileSynthDef("chipsy")
-chipsy.add()
-
-click = FileSynthDef("click")
-click.add()
-
-clip = FileSynthDef("clip")
-clip.add()
-
-cluster = FileSynthDef("cluster")
-cluster.add()
-
-combs = FileSynthDef("combs")
-combs.add()
-
-cs80lead = FileSynthDef("cs80lead")
-cs80lead.add()
-
-dafbass = FileSynthDef("dafbass")
-dafbass.add()
-
-dblbass = FileSynthDef("dblbass")
-dblbass.add()
-
-donkysub = FileSynthDef("donkysub")
-donkysub.add()
-
-donorgan = FileSynthDef("donorgan")
-donorgan.add()
-
-dopple = FileSynthDef("dopple")
-dopple.add()
-
-drone = FileSynthDef("drone")
-drone.add()
-
-dustv = FileSynthDef("dustv")
-dustv.add()
-
-ebass = FileSynthDef("ebass")
-ebass.add()
-
-ecello = FileSynthDef("ecello")
-ecello.add()
-
-eeri = FileSynthDef("eeri")
-eeri.add()
-
-eoboe = FileSynthDef("eoboe")
-eoboe.add()
-
-epiano = FileSynthDef("epiano")
-epiano.add()
-
-faim = FileSynthDef("faim")
-faim.add()
-
-faim2 = FileSynthDef("faim2")
-faim2.add()
-
-fbass = FileSynthDef("fbass")
-fbass.add()
-
-filthysaw = FileSynthDef("filthysaw")
-filthysaw.add()
-
-flute = FileSynthDef("flute")
-flute.add()
-
-fm = FileSynthDef("fm")
-fm.add()
-
-fmbass = FileSynthDef("fmbass")
-fmbass.add()
-
-fmrhodes = FileSynthDef("fmrhodes")
-fmrhodes.add()
-
-garfield = FileSynthDef("garfield")
-garfield.add()
-
-# gaze = FileSynthDef("gaze")
-# gaze.add()
-
-glitchbass = FileSynthDef("glitchbass")
-glitchbass.add()
-
-glitcher = FileSynthDef("glitcher")
-glitcher.add()
-
-grat = FileSynthDef("grat")
-grat.add()
-
-gray = FileSynthDef("gray")
-gray.add()
-
-gsynth = FileSynthDef("gsynth")
-gsynth.add()
-
-harp = FileSynthDef("harp")
-harp.add()
-
-hnoise = FileSynthDef("hnoise")
-hnoise.add()
-
-hoover = FileSynthDef("hoover")
-hoover.add()
-
-kalimba = FileSynthDef("kalimba")
-kalimba.add()
-
-ladder = FileSynthDef("ladder")
-ladder.add()
-
-lapin = FileSynthDef("lapin")
-lapin.add()
-
-laserbeam = FileSynthDef("laserbeam")
-laserbeam.add()
-
-latoo = FileSynthDef("latoo")
-latoo.add()
-
-lazer = FileSynthDef("lazer")
-lazer.add()
-
-lfnoise = FileSynthDef("lfnoise")
-lfnoise.add()
-
-linesaw = FileSynthDef("linesaw")
-linesaw.add()
-
-longsaw = FileSynthDef("longsaw")
-longsaw.add()
-
-mhpad = FileSynthDef("mhpad")
-mhpad.add()
-
-mhping = FileSynthDef("mhping")
-mhping.add()
-
-moogbass = FileSynthDef("moogbass")
-moogbass.add()
-
-moogpluck = FileSynthDef("moogpluck")
-moogpluck.add()
-
-moogpluck2 = FileSynthDef("moogpluck2")
-moogpluck2.add()
-
-mpluck = FileSynthDef("mpluck")
-mpluck.add()
-
-noisynth = FileSynthDef("noisynth")
-noisynth.add()
-
-noquarter = FileSynthDef("noquarter")
-noquarter.add()
-
-organ = FileSynthDef("organ")
-organ.add()
-
-organ2 = FileSynthDef("organ2")
-organ2.add()
-
-pbass = FileSynthDef("pbass")
-pbass.add()
-
-phazer = FileSynthDef("phazer")
-phazer.add()
-
-pianovel = FileSynthDef("pianovel")
-pianovel.add()
-
-pink = FileSynthDef("pink")
-pink.add()
-
-pmcrotal = FileSynthDef("pmcrotal")
-pmcrotal.add()
-
-ppad = FileSynthDef("ppad")
-ppad.add()
-
-prayerbell = FileSynthDef("prayerbell")
-prayerbell.add()
-
-prof = FileSynthDef("prof")
-prof.add()
-
-radio = FileSynthDef("radio")
-radio.add()
-
-rhodes = FileSynthDef("rhodes")
-rhodes.add()
-
-rhpiano = FileSynthDef("rhpiano")
-rhpiano.add()
-
-risseto = FileSynthDef("risseto")
-risseto.add()
-
-rissetobell = FileSynthDef("rissetobell")
-rissetobell.add()
-
-rlead = FileSynthDef("rlead")
-rlead.add()
-
-rsaw = FileSynthDef("rsaw")
-rsaw.add()
-
-rsin = FileSynthDef("rsin")
-rsin.add()
-
-scatter = FileSynthDef("scatter")
-scatter.add()
-
-shore = FileSynthDef("shore")
-shore.add()
-
-sillyvoice = FileSynthDef("sillyvoice")
-sillyvoice.add()
-
-sine = FileSynthDef("sine")
-sine.add()
-
-siren = FileSynthDef("siren")
-siren.add()
-
-spacesaw = FileSynthDef("spacesaw")
-spacesaw.add()
-
-sos = FileSynthDef("sos")
-sos.add()
-
-sosbell = FileSynthDef("sosbell")
-sosbell.add()
-
-spick = FileSynthDef("spick")
-spick.add()
-
-sputter = FileSynthDef("sputter")
-sputter.add()
-
-square = FileSynthDef("square")
-square.add()
-
-ssaw = FileSynthDef("ssaw")
-ssaw.add()
-
-# stargazer = FileSynthDef("stargazer")
-# stargazer.add()
-
-steeldrum = FileSynthDef("steeldrum")
-steeldrum.add()
-
-strings = FileSynthDef("strings")
-strings.add()
-
-subbass = FileSynthDef("subbass")
-subbass.add()
-
-subbass2 = FileSynthDef("subbass2")
-subbass2.add()
-
-supersaw = FileSynthDef("supersaw")
-supersaw.add()
-
-tb303 = FileSynthDef("tb303")
-tb303.add()
-
-total = FileSynthDef("total")
-total.add()
-
-tremsynth = FileSynthDef("tremsynth")
-tremsynth.add()
-
-tribell = FileSynthDef("tribell")
-tribell.add()
-
-tritri = FileSynthDef("tritri")
-tritri.add()
-
-triwave = FileSynthDef("triwave")
-triwave.add()
-
-tubularbell = FileSynthDef("tubularbell")
-tubularbell.add()
-
-tworgan = FileSynthDef("tworgan")
-tworgan.add()
-
-tworgan2 = FileSynthDef("tworgan2")
-tworgan2.add()
-
-tworgan3 = FileSynthDef("tworgan3")
-tworgan3.add()
-
-tworgan4 = FileSynthDef("tworgan4")
-tworgan4.add()
-
-varicelle = FileSynthDef("varicelle")
-varicelle.add()
-
-vibass = FileSynthDef("vibass")
-vibass.add()
-
-vinsine = FileSynthDef("vinsine")
-vinsine.add()
-
-virus = FileSynthDef("virus")
-virus.add()
-
-waves = FileSynthDef("waves")
-waves.add()
-
-windmaker = FileSynthDef("windmaker")
-windmaker.add()
-
-wobble = FileSynthDef("wobble")
-wobble.add()
-
-wobblebass = FileSynthDef("wobblebass")
-wobblebass.add()
-
-wsaw = FileSynthDef("wsaw")
-wsaw.add()
-
-wsawbass = FileSynthDef("wsawbass")
-wsawbass.add()
-
-xylophone = FileSynthDef("xylophone")
-xylophone.add()
+#sawbass = FileSynthDef('sawbass')
+#sawbass.add()
+#
+#prophet = FileSynthDef('prophet')
+#prophet.add()
+#
+#pads = FileSynthDef('pads')
+#pads.add()
+#
+#pasha = FileSynthDef('pasha')
+#pasha.add()
+#
+#ambi = FileSynthDef("ambi")
+#ambi.add()
+#
+#space = FileSynthDef("space")
+#space.add()
+#
+#keys = FileSynthDef("keys")
+#keys.add()
+#
+#bass = FileSynthDef("bass")
+#bass.add()
+#
+#dbass = FileSynthDef("dbass")
+#dbass.add()
+#
+#sinepad = FileSynthDef("sinepad")
+#sinepad.add()
+#
+#video = FileSynthDef("video")
+#video.add()
+#
+#viola = FileSynthDef("viola")
+#viola.add()
+#
+#hydra = FileSynthDef("hydra")
+#hydra.add()
+#
+#abass = FileSynthDef("abass")
+#abass.add()
+#
+#acidbass = FileSynthDef("acidbass")
+#acidbass.add()
+#
+#alva = FileSynthDef("alva")
+#alva.add()
+#
+#angel = FileSynthDef("angel")
+#angel.add()
+#
+#angst = FileSynthDef("angst")
+#angst.add()
+#
+#bassguitar = FileSynthDef("bassguitar")
+#bassguitar.add()
+#
+#bbass = FileSynthDef("bbass")
+#bbass.add()
+#
+#bchaos = FileSynthDef("bchaos")
+#bchaos.add()
+#
+#bellmod = FileSynthDef("bellmod")
+#bellmod.add()
+#
+#benoit = FileSynthDef("benoit")
+#benoit.add()
+#
+#birdy = FileSynthDef("birdy")
+#birdy.add()
+#
+#blips = FileSynthDef("blips")
+#blips.add()
+#
+#bnoise = FileSynthDef("bnoise")
+#bnoise.add()
+#
+#borgan = FileSynthDef("borgan")
+#borgan.add()
+#
+#bounce = FileSynthDef("bounce")
+#bounce.add()
+#
+#bphase = FileSynthDef("bphase")
+#bphase.add()
+#
+#brass = FileSynthDef("brass")
+#brass.add()
+#
+#brown = FileSynthDef("brown")
+#brown.add()
+#
+#chimebell = FileSynthDef("chimebell")
+#chimebell.add()
+#
+#chipsy = FileSynthDef("chipsy")
+#chipsy.add()
+#
+#click = FileSynthDef("click")
+#click.add()
+#
+#clip = FileSynthDef("clip")
+#clip.add()
+#
+#cluster = FileSynthDef("cluster")
+#cluster.add()
+#
+#combs = FileSynthDef("combs")
+#combs.add()
+#
+#cs80lead = FileSynthDef("cs80lead")
+#cs80lead.add()
+#
+#dafbass = FileSynthDef("dafbass")
+#dafbass.add()
+#
+#dblbass = FileSynthDef("dblbass")
+#dblbass.add()
+#
+#donkysub = FileSynthDef("donkysub")
+#donkysub.add()
+#
+#donorgan = FileSynthDef("donorgan")
+#donorgan.add()
+#
+#dopple = FileSynthDef("dopple")
+#dopple.add()
+#
+#drone = FileSynthDef("drone")
+#drone.add()
+#
+#dustv = FileSynthDef("dustv")
+#dustv.add()
+#
+#ebass = FileSynthDef("ebass")
+#ebass.add()
+#
+#ecello = FileSynthDef("ecello")
+#ecello.add()
+#
+#eeri = FileSynthDef("eeri")
+#eeri.add()
+#
+#eoboe = FileSynthDef("eoboe")
+#eoboe.add()
+#
+#epiano = FileSynthDef("epiano")
+#epiano.add()
+#
+#faim = FileSynthDef("faim")
+#faim.add()
+#
+#faim2 = FileSynthDef("faim2")
+#faim2.add()
+#
+#fbass = FileSynthDef("fbass")
+#fbass.add()
+#
+#filthysaw = FileSynthDef("filthysaw")
+#filthysaw.add()
+#
+#flute = FileSynthDef("flute")
+#flute.add()
+#
+#fm = FileSynthDef("fm")
+#fm.add()
+#
+#fmbass = FileSynthDef("fmbass")
+#fmbass.add()
+#
+#fmrhodes = FileSynthDef("fmrhodes")
+#fmrhodes.add()
+#
+#garfield = FileSynthDef("garfield")
+#garfield.add()
+#
+## gaze = FileSynthDef("gaze")
+## gaze.add()
+#
+#glitchbass = FileSynthDef("glitchbass")
+#glitchbass.add()
+#
+#glitcher = FileSynthDef("glitcher")
+#glitcher.add()
+#
+#grat = FileSynthDef("grat")
+#grat.add()
+#
+#gray = FileSynthDef("gray")
+#gray.add()
+#
+#gsynth = FileSynthDef("gsynth")
+#gsynth.add()
+#
+#harp = FileSynthDef("harp")
+#harp.add()
+#
+#hnoise = FileSynthDef("hnoise")
+#hnoise.add()
+#
+#hoover = FileSynthDef("hoover")
+#hoover.add()
+#
+#kalimba = FileSynthDef("kalimba")
+#kalimba.add()
+#
+#ladder = FileSynthDef("ladder")
+#ladder.add()
+#
+#lapin = FileSynthDef("lapin")
+#lapin.add()
+#
+#laserbeam = FileSynthDef("laserbeam")
+#laserbeam.add()
+#
+#latoo = FileSynthDef("latoo")
+#latoo.add()
+#
+#lazer = FileSynthDef("lazer")
+#lazer.add()
+#
+#lfnoise = FileSynthDef("lfnoise")
+#lfnoise.add()
+#
+#linesaw = FileSynthDef("linesaw")
+#linesaw.add()
+#
+#longsaw = FileSynthDef("longsaw")
+#longsaw.add()
+#
+#mhpad = FileSynthDef("mhpad")
+#mhpad.add()
+#
+#mhping = FileSynthDef("mhping")
+#mhping.add()
+#
+#moogbass = FileSynthDef("moogbass")
+#moogbass.add()
+#
+#moogpluck = FileSynthDef("moogpluck")
+#moogpluck.add()
+#
+#moogpluck2 = FileSynthDef("moogpluck2")
+#moogpluck2.add()
+#
+#mpluck = FileSynthDef("mpluck")
+#mpluck.add()
+#
+#noisynth = FileSynthDef("noisynth")
+#noisynth.add()
+#
+#noquarter = FileSynthDef("noquarter")
+#noquarter.add()
+#
+#organ = FileSynthDef("organ")
+#organ.add()
+#
+#organ2 = FileSynthDef("organ2")
+#organ2.add()
+#
+#pbass = FileSynthDef("pbass")
+#pbass.add()
+#
+#phazer = FileSynthDef("phazer")
+#phazer.add()
+#
+#pianovel = FileSynthDef("pianovel")
+#pianovel.add()
+#
+#pink = FileSynthDef("pink")
+#pink.add()
+#
+#pmcrotal = FileSynthDef("pmcrotal")
+#pmcrotal.add()
+#
+#ppad = FileSynthDef("ppad")
+#ppad.add()
+#
+#prayerbell = FileSynthDef("prayerbell")
+#prayerbell.add()
+#
+#prof = FileSynthDef("prof")
+#prof.add()
+#
+#radio = FileSynthDef("radio")
+#radio.add()
+#
+#rhodes = FileSynthDef("rhodes")
+#rhodes.add()
+#
+#rhpiano = FileSynthDef("rhpiano")
+#rhpiano.add()
+#
+#risseto = FileSynthDef("risseto")
+#risseto.add()
+#
+#rissetobell = FileSynthDef("rissetobell")
+#rissetobell.add()
+#
+#rlead = FileSynthDef("rlead")
+#rlead.add()
+#
+#rsaw = FileSynthDef("rsaw")
+#rsaw.add()
+#
+#rsin = FileSynthDef("rsin")
+#rsin.add()
+#
+#scatter = FileSynthDef("scatter")
+#scatter.add()
+#
+#shore = FileSynthDef("shore")
+#shore.add()
+#
+#sillyvoice = FileSynthDef("sillyvoice")
+#sillyvoice.add()
+#
+#sine = FileSynthDef("sine")
+#sine.add()
+#
+#siren = FileSynthDef("siren")
+#siren.add()
+#
+#spacesaw = FileSynthDef("spacesaw")
+#spacesaw.add()
+#
+#sos = FileSynthDef("sos")
+#sos.add()
+#
+#sosbell = FileSynthDef("sosbell")
+#sosbell.add()
+#
+#spick = FileSynthDef("spick")
+#spick.add()
+#
+#sputter = FileSynthDef("sputter")
+#sputter.add()
+#
+#square = FileSynthDef("square")
+#square.add()
+#
+#ssaw = FileSynthDef("ssaw")
+#ssaw.add()
+#
+## stargazer = FileSynthDef("stargazer")
+## stargazer.add()
+#
+#steeldrum = FileSynthDef("steeldrum")
+#steeldrum.add()
+#
+#strings = FileSynthDef("strings")
+#strings.add()
+#
+#subbass = FileSynthDef("subbass")
+#subbass.add()
+#
+#subbass2 = FileSynthDef("subbass2")
+#subbass2.add()
+#
+#supersaw = FileSynthDef("supersaw")
+#supersaw.add()
+#
+#tb303 = FileSynthDef("tb303")
+#tb303.add()
+#
+#total = FileSynthDef("total")
+#total.add()
+#
+#tremsynth = FileSynthDef("tremsynth")
+#tremsynth.add()
+#
+#tribell = FileSynthDef("tribell")
+#tribell.add()
+#
+#tritri = FileSynthDef("tritri")
+#tritri.add()
+#
+#triwave = FileSynthDef("triwave")
+#triwave.add()
+#
+#tubularbell = FileSynthDef("tubularbell")
+#tubularbell.add()
+#
+#tworgan = FileSynthDef("tworgan")
+#tworgan.add()
+#
+#tworgan2 = FileSynthDef("tworgan2")
+#tworgan2.add()
+#
+#tworgan3 = FileSynthDef("tworgan3")
+#tworgan3.add()
+#
+#tworgan4 = FileSynthDef("tworgan4")
+#tworgan4.add()
+#
+#varicelle = FileSynthDef("varicelle")
+#varicelle.add()
+#
+#vibass = FileSynthDef("vibass")
+#vibass.add()
+#
+#vinsine = FileSynthDef("vinsine")
+#vinsine.add()
+#
+#virus = FileSynthDef("virus")
+#virus.add()
+#
+#waves = FileSynthDef("waves")
+#waves.add()
+#
+#windmaker = FileSynthDef("windmaker")
+#windmaker.add()
+#
+#wobble = FileSynthDef("wobble")
+#wobble.add()
+#
+#wobblebass = FileSynthDef("wobblebass")
+#wobblebass.add()
+#
+#wsaw = FileSynthDef("wsaw")
+#wsaw.add()
+#
+#wsawbass = FileSynthDef("wsawbass")
+#wsawbass.add()
+#
+#xylophone = FileSynthDef("xylophone")
+#xylophone.add()
