+++ renardo_lib/renardo_lib/runtime/synthdefs_initialisation/special_synthdefs.py
@@ -1,19 +1,19 @@
-from renardo_lib.SCLang.SCLangExperimentalBindings import *
-from renardo_lib.SCLang.SynthDef import DefaultSynthDef
-from renardo_lib.SCLang.SampleSynthDefs import SampleSynthDef, LoopSynthDef, StretchSynthDef, GranularSynthDef
-from renardo_lib.SCLang.SCLangExperimentalBindings import Env
+from renardo_lib.SynthDefManagement.SCLangExperimentalPythonBindings import *
+from renardo_lib.SynthDefManagement.PySynthDef import DefaultSynthDef
+from renardo_lib.SynthDefManagement.SampleSynthDefs import SampleSynthDef, LoopSynthDef, StretchSynthDef, GranularSynthDef
+from renardo_lib.SynthDefManagement.SCLangExperimentalPythonBindings import Env
 
 # Sample Player
 
-with SampleSynthDef("play1") as play:
-    play.osc = PlayBuf.ar(1, play.buf, BufRateScale.ir(
-        play.buf) * play.rate, startPos=BufSampleRate.kr(play.buf) * play.pos)
-    play.osc = play.osc * play.amp
-
-with SampleSynthDef("play2") as play:
-    play.osc = PlayBuf.ar(2, play.buf, BufRateScale.ir(
-        play.buf) * play.rate, startPos=BufSampleRate.kr(play.buf) * play.pos)
-    play.osc = play.osc * play.amp
+#with SampleSynthDef("play1") as play:
+#    play.osc = PlayBuf.ar(1, play.buf, BufRateScale.ir(
+#        play.buf) * play.rate, startPos=BufSampleRate.kr(play.buf) * play.pos)
+#    play.osc = play.osc * play.amp
+#
+#with SampleSynthDef("play2") as play:
+#    play.osc = PlayBuf.ar(2, play.buf, BufRateScale.ir(
+#        play.buf) * play.rate, startPos=BufSampleRate.kr(play.buf) * play.pos)
+#    play.osc = play.osc * play.amp
 
 
 loop = LoopSynthDef()
